Moved definitions to registry
